requirements.txt
requests>=2.32.3
dash>=2.18.1
dash-bootstrap-components>=1.6.1rc2
Flask>=3.0.3
gunicorn>=20.1.0
gevent >=21.12.0
pytz>=2024.2
python-dotenv>=1.0.1
diskcache>=5.6.3
Flask-Caching>=2.3.0
pandas>=2.2.3